Allow loop64 tail blending to evict previously added fragments
Previously, tail blending could happen only if the early test against the furthest fragment had failed.

That had a few logical issues that resulted in subpar quality:

- Only new fragments could be sent to tail blending, but previously added ones were evicted into the void.
- The early test is an unsynchronized conservative check. Even if it has passed, it can silently fail at any moment later.

Now, the furthest fragment, which is sent to one of the shader invocations as the previous value from `atomicMin`, goes to tail blending instead.

This gives a result that is largely identical to what the spinlock and the unordered interlock methods offer.

Signed-off-by: Triang3l <triang3l@yandex.ru>